Fix workflow executor line-limit check
