+++ AGENTS.md
@@ -123,7 +123,7 @@ An expression is only shipped when every one of these is done in the same change
 4. **Update the DSL prompt and the expression metadata** (`workflow_dsl_prompt.py`, the expression dialog's method list and autocomplete previews) so the AI assistant and the dialog can offer it.
 5. **Update the docs** — the expression reference under `frontend/src/docs/content/` — plus a release tour entry if the expression is user-visible.
 
-**Names Python reserves are a trap.** `$global` was silently broken for months because `global.x` cannot be parsed by `ast.parse`: every `$global…` expression fell through to the fallback resolver, which only knows a handful of no-argument string helpers, and returned `null` for everything else. `alias_reserved_context_names()` in `workflow_executor.py` rewrites such names before parsing, and every parse must go through `_parse_expression_tree` / `HeymExpressionEval.eval` so the rewrite is applied. If you add a context root, check it is not a Python keyword.
+**Names Python reserves are a trap.** `$global` was silently broken for months because `global.x` cannot be parsed by `ast.parse`: every `$global…` expression fell through to the fallback resolver, which only knows a handful of no-argument string helpers, and returned `null` for everything else. `alias_reserved_context_names()` in `expression_syntax.py` rewrites such names before parsing, and every parse must go through `_parse_expression_tree` / `HeymExpressionEval.eval` so the rewrite is applied. If you add a context root, check it is not a Python keyword.
 
 ### OpenTelemetry tracing (keep span seams aligned)
 OTel tracing is env-gated (`HEYM_OTEL_ENABLED`, disabled by default) and bootstrapped in `backend/app/observability/tracing.py` from `app/main.py`'s `setup_tracing(app)`. Spans are added at three seams only:
